requirements.txt
alembic==1.9.4
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
PyJWT==2.6.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
